+++ c++/provider/DHParameterGenerator.h
@@ -37,7 +37,7 @@ namespace beecrypt {
 		class DHParameterGenerator : public beecrypt::security::AlgorithmParameterGeneratorSpi
 		{
 		private:
-			size_t _size;
+			int _size;
 			DHParameterSpec* _spec;
 			SecureRandom* _srng;
 
@@ -45,7 +45,7 @@ namespace beecrypt {
 			virtual AlgorithmParameters* engineGenerateParameters();
 
 			virtual void engineInit(const AlgorithmParameterSpec&, SecureRandom*) throw (InvalidAlgorithmParameterException);
-			virtual void engineInit(size_t, SecureRandom*) throw (InvalidParameterException);
+			virtual void engineInit(int, SecureRandom*) throw (InvalidParameterException);
 
 		public:
 			DHParameterGenerator();
